+++ oasyce_sdk/samantha/__main__.py
@@ -1,4 +1,0 @@
-"""Allow running as: python -m oasyce_sdk.samantha"""
-from .server import main
-
-main()